move health config syncer impl to alerting package
